+++ Source/WebSites/DevelopmentWebsite/UserControl/TributeHeader.ascx.cs
@@ -97,7 +97,7 @@ public partial class UserControl_TributeHeader : System.Web.UI.UserControl
             else if (intUserType == 2)
             {
                 _userName = objSessionvalue.UserName;
-                int NetCreditPoints;
+                double NetCreditPoints;
                 UserRegistration _objUserReg = new UserRegistration();
                 Users objUsers = new Users();
                 objUsers.UserId = objSessionvalue.UserId;
